cast timeout to float instead of int

+++ in_toto/runlib.py
@@ -358,11 +358,11 @@ def execute_link(link_cmd_args, record_streams):
   if record_streams:
     return_code, stdout_str, stderr_str = \
         securesystemslib.process.run_duplicate_streams(link_cmd_args,
-            timeout=int(in_toto.settings.LINK_CMD_EXEC_TIMEOUT))
+            timeout=float(in_toto.settings.LINK_CMD_EXEC_TIMEOUT))
 
   else:
     process = securesystemslib.process.run(link_cmd_args, check=False,
-      timeout=int(in_toto.settings.LINK_CMD_EXEC_TIMEOUT),
+      timeout=float(in_toto.settings.LINK_CMD_EXEC_TIMEOUT),
       stdout=securesystemslib.process.DEVNULL,
       stderr=securesystemslib.process.DEVNULL)
     stdout_str = stderr_str = ""
